+++ src/StatePulse.NET/IStateAccessor.cs
@@ -6,7 +6,7 @@ public delegate void OnChangeEventHandler<in T>(object sender, T args);
 public interface IStateAccessor<out TState>
 {
     TState State { get; }
-    event OnChangeEventHandler<TState>? StateChanged;
-    event EventHandler? StateChangedNoDetails;
+    event OnChangeEventHandler<TState>? OnStateChanged;
+    event EventHandler? OnStateChangedNoDetails;
 
 }